# [Strings] Split String Optimizations