--- a/common/string_util.h
+++ b/common/string_util.h
@@ -176,7 +176,7 @@ std::vector<std::string> join_tuple(const std::string &glue, const std::pair<cha
 	return output;
 }
 
-std::vector<std::string> SplitString(const std::string &s, char delim);
+std::vector<std::string> SplitString(const std::string &s, const char delim = ',');
 std::string::size_type search_deliminated_string(const std::string &haystack, const std::string &needle, const char deliminator = ',');
 std::string EscapeString(const char *src, size_t sz);
 std::string EscapeString(const std::string &s);