Some clean up since we had duplicate functions and optimizations

This is on a VERY long SpecialAbilities (taken from the PEQ's DB)
 
```
    --------------------------------------------------------------
    Benchmark                       Time           CPU Iterations
    --------------------------------------------------------------
    bench_oldsplit               5201 ns       5201 ns     129500
    bench_split                  1269 ns       1269 ns     548906
```